fix(stats): banc de stats ressuscité, mécanismes orphelins re-branchés
La purge de la démo 2D avait orphelin le time-skip et une partie du
pipeline stats. Restauré et rendu transversal :
- waitGameTime/sleepGameTime routés par advanceGameTime (le lit soigne
  à nouveau) ; guérison des blessures gated sur le repos ; helpers
  transversaux sleepCharacter/waitCharacter (CharacterTick).
- Buildups d'arme re-branchés en 3D : DamageEvent porte
  buildupType/buildupAmount (remplis par weaponDamageEvent), appliqués
  dans resolveStrikeDamage sur les coups passés ; rollInjury re-câblé
  dans la frappe (canal physique dominant, Torso v1) ; consommables à
  buildup no-op corrigés (UiRouter).
- CharacterStatsPanel (F7) et StatsScene restaurés dans le menu démo ;
  LandscapeScene ne fait qu'utiliser l'état du joueur (applySleep/
  applyWait, combatRng vers Player/ProjectileContext).
- Audit de conformité STATS.md : docs/AUDIT/STATS-CONFORMANCE.md
  (19 décisions design à trancher, doc en retard listé).

Le fichier LandscapeScene porte aussi sa part des étapes 2 (updateWarmup
en fin d'update) et 6 (commentaires/logs) de la revue.

+++ game/scenes/CombatArenaScene.cpp
@@ -180,8 +180,8 @@ void CombatArenaScene::onEnter() {
         sparks.colorStart = { 1.0f, 0.9f, 0.35f, 1.0f };
         sparks.colorEnd = { 1.0f, 0.25f, 0.05f, 0.0f };
         // Cosmetic seed: position hash + running count (free RNG, §8).
-        const u32 seed = static_cast<u32>(event.position.x * 73.0f) ^
-                         (static_cast<u32>(event.position.y * 179.0f) << 8) ^
+        const u32 seed = static_cast<u32>(static_cast<i32>(event.position.x * 73.0f)) ^
+                         (static_cast<u32>(static_cast<i32>(event.position.y * 179.0f)) << 8) ^
                          particles.count();
         particles.spawn(sparks, event.position, seed);
     });
